add image picker

+++ Countdown/Countdown WatchKit Extension/Utilities/Gallery.h
@@ -9,6 +9,6 @@
 #import <Foundation/Foundation.h>
 
 @interface Gallery : NSObject
--(NSString*) imageForRow: (int) row;
+-(NSString*) imageForRow: (NSUInteger) row;
 -(NSUInteger) numberOfRows;
 @end
